svn path=/trunk/; revision=69

+++ ibaDatCoordinator/Processing/FileQuotaCleanup.cs
@@ -25,7 +25,15 @@ namespace iba.Processing
 
         public void AddFile(string filename)
         {
-            throw new Exception("The method or operation is not implemented.");
+            try
+            {
+                FileInfo inf = new FileInfo(filename);
+                m_size += (ulong)inf.Length;
+                m_files.Add(filename);
+            }
+            catch
+            {
+            }
         }
 
         public void ResetTask(TaskDataUNC newTask, string extension)
@@ -68,18 +76,45 @@ namespace iba.Processing
             }
         }
 
-        public void Clean()
+        private void Log(Logging.Level level, string message, string datfile)
         {
-            while (m_size > m_task.Quota && m_files.Count > 0)
+            if (LogData.Data.Logger.IsOpen)
+            {
+                LogExtraData data = new LogExtraData(datfile, m_task, m_task.ParentConfigurationData);
+                LogData.Data.Logger.Log(level, message, (object)data);
+            }
+        }
+
+        public void Clean(string datfile)
+        {
+            while (m_size > (m_task.Quota * 1024 * 1024) && m_files.Count > 0)
             {
                 string file = m_files[0];
                 try
                 {
+                    if (!File.Exists(file))
+                    {
+                        Reset();
+                        continue;
+                    }
                     string origPath = (new DirectoryInfo(m_task.DestinationMapUNC)).FullName;
                     FileInfo inf = new FileInfo(file);
                     DirectoryInfo parent = inf.Directory;
                     ulong size = (ulong) inf.Length;
-                    File.Delete(file);
+                    try
+                    {
+
+                        File.Delete(file);
+                        string message = String.Format(iba.Properties.Resources.logCleanupSuccessRemoveFile, file);
+                        Log(iba.Logging.Level.Info, message, datfile);
+
+                    }
+                    catch (Exception ex)
+                    {
+                        string message = String.Format(iba.Properties.Resources.logCleanupErrorRemoveFile, file, ex.Message);
+                        Log(iba.Logging.Level.Warning, message, datfile);
+                        throw ex;
+                    }
                     if (size > m_size)
                         m_size = 0;
                     else
@@ -87,7 +122,18 @@ namespace iba.Processing
                     while (parent.GetDirectories().Length == 0 && parent.GetFiles().Length == 0 && parent.FullName != origPath)
                     {
                         DirectoryInfo grandParent = parent.Parent;
-                        Directory.Delete(parent.FullName);
+                        try
+                        {
+                            Directory.Delete(parent.FullName);
+                            string message = String.Format(iba.Properties.Resources.logCleanupSuccessRemoveEmptyDirectory, parent.FullName);
+                            Log(iba.Logging.Level.Info, message, datfile);
+                        }
+                        catch (Exception ex)
+                        {
+                            string message = String.Format(iba.Properties.Resources.logCleanupErrorRemoveDirectory, parent.FullName, ex.Message);
+                            Log(iba.Logging.Level.Warning, message, datfile);
+                            throw ex;
+                        }
                         parent = grandParent;
                     }
                 }
